+++ aidr-tagger-api/src/main/java/qa/qcri/aidr/predictui/api/ModelResource.java
@@ -62,7 +62,7 @@ public class ModelResource {
    @Produces(MediaType.APPLICATION_JSON)
    @Path("crisis/{crisisID}")
    public Response getModelByCrisisID(@PathParam("crisisID") int crisisID){
-       List<ModelWrapper> modelList = modelLocalEJB.getModelByCrisisID(crisisID);
+       List<ModelWrapper> modelList = modelLocalEJB.getModelByCrisisID((long) crisisID);
        ResponseWrapper response = new ResponseWrapper();
        if (modelList == null){
            response.setMessage("No models found for the given crisis-id");
